requirements.txt
streamlit
tensorflow
Pillow
numpy
flask